requirements_streamlit.txt
streamlit
pillow
opencv-python
numpy
matplotlib